(collect_iterators): Cast tree code to int for indexing.
From-SVN: r3913

+++ gcc/c-iterate.c
@@ -223,7 +223,7 @@ collect_iterators (exp, list)
 	case 'e':
 	case 'r':
 	  {
-	    int num_args = tree_code_length[TREE_CODE (exp)];
+	    int num_args = tree_code_length[(int) TREE_CODE (exp)];
 	    int i;
 
 	    /* Some tree codes have RTL, not trees, as operands.  */
